BlockManager Behaviour Update + alt KOTH
BlockManager now defaults its init and behaviour if one is not assigned.
A default BlockManager is now created if there is not one in a level. This manager collects the loose statics, generates spawn platforms, and creates no blocks.

An alternate KOTH has been created (but is not yet finished).

A few touch-ups were added to the inspector view of code (i.e. spacing and hiding unnecessary fields).